Perbaikan Cuti, Attendance, Ijin, Penambahan security dan setup admin

+++ src/routes/(protected)/api/cuti/[id]/delete/+server.ts
@@ -1,16 +1,39 @@
 import { error, json } from "@sveltejs/kit";
-import { prisma, prismaErrorHandler } from '@lib/utils.js'
+import { pecahArray, prisma, prismaErrorHandler } from '@lib/utils.js'
+import { unlink } from 'fs/promises'
+import { existsSync } from 'fs'
+import path from 'path'
 
-export async function DELETE({params}){
+export async function DELETE({ params, locals }) {
     try {
-        const {id} = params
-        await prisma.cuti.delete({
-            where:{
-                cuti_id:id,
+        const { id } = params
+        const { userProfile } = locals
+
+        if (!pecahArray(userProfile.access_cuti, "D")) throw new Error("Tidak dapat menghapus cuti, anda tidak memiliki akses")
+
+        const [getCuti] = await prisma.$queryRawUnsafe(`
+            SELECT c.attachment, c.cuti_group_id,
+                (SELECT CAST(COUNT(cuti_id) as UNSIGNED) 
+                FROM cuti WHERE cuti_group_id = c.cuti_group_id) AS jumlah
+            FROM cuti c
+            WHERE c.cuti_id = ?`, id) as { attachment: string, cuti_group_id: string, jumlah: number }[]
+
+        const filePath = path.resolve(process.env.ATTACH_CUTI + getCuti.attachment)
+        const fileExist = existsSync(filePath)
+
+
+        const deleteCuti = await prisma.cuti.delete({
+            where: {
+                cuti_id: id,
                 status: "Waiting",
             }
         })
-        return json({message:"Data successfully deleted"})
+
+        if (deleteCuti && fileExist && getCuti.jumlah == 1) {
+            await unlink(filePath);
+        }
+
+        return json({ message: "Data successfully deleted" })
     } catch (err) {
         error(500, prismaErrorHandler(err))
     }
